requirements.txt
fastapi
uvicorn
jinja2
python-multipart
firebase-admin
google-cloud-firestore
google-cloud-storage
httpx
pydantic
python-jose[cryptography]
PyJWT
passlib[bcrypt]
python-csv
Pillow